# openapi.yaml
openapi: 3.0.0
paths:
...
  /tenant:
    get:
      operationId: runHello
      summary: ""
      description: Say hello only
      parameters: []
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal error
      tags: &a1
        - TENANT
    post:
      operationId: runCreate
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tenant"
      responses:
        "201":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a1
...
  /profile:
    get:
      operationId: getProfile
      summary: ""
      description: Get current user profile
      parameters: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyProfileApiSchema"
        "401":
          description: Undefine profile
          content:
            application/json:
              schema:
                type: object
      tags: &a12
        - PROFILE
  /profile/session:
    get:
      operationId: getSession
      summary: ""
      description: Get current user session
      parameters: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/String"
        "401":
          description: Expired
          content:
            application/json:
              schema:
                type: object
      tags: *a12
  /profile/tenants:
    get:
      operationId: getAllTenants
      summary: ""
      description: Get all tenants
      parameters: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/"
        "401":
          description: Undefine profile
          content:
            application/json:
              schema:
                type: object
      tags: *a12
  /profile/tenant:
    post:
      operationId: createTenant
      summary: ""
      description: create new tenant
      parameters: []
      requestBody:
        required: true
        description: create tenant name
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegTenantApiSchema"
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                type: object
      tags: *a12
  /profile/invitation/{id}/{decision}:
    post:
      operationId: decideInvitation
      summary: ""
      description: accept or reject invitation to tenant
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: decision
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                type: object
      tags: *a12
...
info:
  title: SimpleApp Demo1
  description: Try CRUD
  contact: {}
...
servers: []
components:
  securitySchemes:
    x-org:
      type: apiKey
      in: header
      name: x-org
      description: "base 64 url encode. example: MC0wLTA/MS0xLTE"
    oauth2:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: /protocol/openid-connect/auth
          scopes: {}
  schemas:
...
    TenantOwner:
      type: object
      properties:
        _id:
          type: string
          default: ""
        label:
          type: string
          default: ""
        uid:
          type: string
          default: ""
      required:
        - _id
        - label
        - uid
    Tenant:
      type: object
      properties:
        _id:
          type: string
          format: uuid
          default: ""
        created:
          type: string
          default: ""
        updated:
          type: string
          default: ""
        createdBy:
          type: string
          default: ""
        updatedBy:
          type: string
          default: ""
        tenantId:
          type: number
          default: 1
        orgId:
          type: number
          default: 1
        branchId:
          type: number
          default: 1
        tenantName:
          type: string
          default: ""
        active:
          type: boolean
          examples:
            - true
          default: true
        description:
          type: string
          default: ""
        owner:
          $ref: "#/components/schemas/TenantOwner"
      required:
        - tenantId
        - tenantName
        - owner
...
    MyProfileApiSchemaBranches:
      type: object
      properties:
        _id:
          type: string
        tenantId:
          type: number
        orgId:
          type: number
        branchId:
          type: number
        group:
          type: string
        xOrg:
          type: string
      required:
        - _id
        - tenantId
        - orgId
        - branchId
        - group
        - xOrg
    MyProfileApiSchemaInvites:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
        created:
          type: string
      required:
        - _id
        - email
        - created
    MyProfileApiSchema:
      type: object
      properties:
        _id:
          type: string
        tenantId:
          type: number
        orgId:
          type: number
        branchId:
          type: number
        email:
          type: string
        uid:
          type: number
        fullName:
          type: number
        group:
          type: number
        roles:
          type: array
          items:
            type: string
        branches:
          type: array
          items:
            $ref: "#/components/schemas/MyProfileApiSchemaBranches"
        invites:
          type: array
          items:
            $ref: "#/components/schemas/MyProfileApiSchemaInvites"
        time:
          type: string
      required:
        - _id
        - tenantId
        - orgId
        - branchId
        - email
        - uid
        - fullName
        - group
        - roles
        - branches
        - invites
        - time
    String:
      type: object
      properties: {}
    RegTenantApiSchema:
      type: object
      properties:
        tenantName:
          type: string
          examples:
            - my company 1
          default: ""
      required:
        - tenantName
...
security:
  - x-org: []
  - oauth2: []